Bot ECAM
ECAM's bot discord ensures security on the server. This bot verifies the identity of students entering the server. It contains a wide variety of functions, including a mute function, a message deleting function and an emergency deleting function, which deletes the messages of someone who has been hacked. There's even more to discover.